# src/main/resources/swagger/api.yml
openapi: 3.0.3
info:
  title: Jasper KMS
  description: REST API for Jasper Knowledge Management Server
  version: '1.1.0'
  license:
    name: MIT License
    url: 'https://github.com/cjmalloy/jasper/blob/master/LICENSE'
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/cjmalloy/Jasper/1.1.0
paths:
...
  /api/v1/ref:
    get:
      tags:
        - Ref
      operationId: getRef
      parameters:
        - name: url
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/url'
        - name: origin
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/origin'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Ref'
        '304':
          description: Not Modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Ref
      operationId: updateRef
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ref'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Ref
      operationId: createRef
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ref'
        required: true
      responses:
        '201':
          description: Created
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ForeignWrite'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags:
        - Ref
      operationId: deleteRef
      parameters:
        - name: url
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/url'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags:
        - Ref
      operationId: patchRef
      parameters:
        - name: url
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/url'
        - name: origin
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/origin'
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: 'https://json.schemastore.org/json-patch'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/ref/list:
    get:
      tags:
        - Ref
      operationId: getRefList
      parameters:
        - name: urls
          in: query
          required: true
          schema:
            maxItems: 100
            minItems: 0
            type: array
            items:
              $ref: '#/components/schemas/url'
        - name: origin
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/origin'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ref'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/ref/page:
    get:
      tags:
        - Ref
      operationId: getRefPage
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 10
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          schema:
            type: array
            items:
              type: string
        - name: query
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/query'
        - name: url
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/url'
        - name: sources
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/url'
        - name: responses
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/url'
        - name: uncited
          in: query
          required: false
          schema:
            type: boolean
        - name: unsourced
          in: query
          required: false
          schema:
            type: boolean
        - name: modifiedAfter
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: pluginResponse
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/plugin-tag'
        - name: noPluginResponse
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/plugin-tag'
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageRef'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/ref/count:
    get:
      tags:
        - Ref
      operationId: countRefs
      parameters:
        - name: query
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/query'
        - name: url
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/url'
        - name: sources
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/url'
        - name: responses
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/url'
        - name: uncited
          in: query
          required: false
          schema:
            type: boolean
        - name: unsourced
          in: query
          required: false
          schema:
            type: boolean
        - name: modifiedAfter
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: pluginResponse
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/plugin-tag'
        - name: noPluginResponse
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/plugin-tag'
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  schemas:
...
    url:
      type: string
      format: uri
      pattern: '([^:/?#]+):(\/\/([^\/?#]*))?([^?#]*)(\?([^#]*))?(#(.*))?'
      description: URL
      example: http://jasper-kms.info/
      maxLength: 4096
    title:
      type: string
      description: Title
      example: Title
      maxLength: 512
    origin:
      type: string
      pattern: '@[a-z]+(\.[a-z])*'
      default: ''
      description: Origin
      example: '@origin'
      maxLength: 64
...
    tag:
      type: string
      pattern: '[_+]?[a-z]+(/[a-z]+)*'
      description: Tag
      example: 'tag'
      maxLength: 64
...
    query:
      type: string
      description: Tag Query
      pattern: '(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))|\(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))([ |]!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*)))*\))([ |:&](!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))|\(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))([ |]!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*)))*\)))*'
      example: '(politics|science):!funny'
      maxLength: 512
...
    plugin-tag:
      type: string
      pattern: '[_+]?plugin/[a-z]+(/[a-z]+)*'
      description: Plugin Tag
      example: 'plugin/thumbnail'
      maxLength: 64
...
    Ref:
      required:
        - published
        - url
      type: object
      properties:
        url:
          $ref: '#/components/schemas/url'
        origin:
          $ref: '#/components/schemas/origin'
        title:
          $ref: '#/components/schemas/title'
        comment:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/tag'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/url'
        alternateUrls:
          type: array
          items:
            $ref: '#/components/schemas/url'
        plugins:
          type: object
          example:
            'plugin/thumbnail':
              url: http://example-thumbnail.com/image.png
        metadata:
          $ref: '#/components/schemas/Metadata'
        published:
          type: string
          format: date-time
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
    Metadata:
      type: object
      properties:
        responses:
          type: array
          items:
            $ref: '#/components/schemas/url'
        internalResponses:
          type: array
          items:
            $ref: '#/components/schemas/url'
        plugins:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/url'
          example:
            'plugin/comment':
              - comment:671ceb54-3fac-4f14-b5e0-3b7270fa1787
              - comment:d4a83102-a94b-4bfb-8914-1863ac673bde
...
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageRef:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Ref'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
    NotFound:
      description: Not Found
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
    ForeignWrite:
      description: Foreign Write
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
    Conflict:
      description: Conflict
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
    Forbidden:
      description: Forbidden
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'